Fix parallel UI to undo the aggressive variant & commit formatting
9188cac05ddfe065783c572260d5c07b8e did not fix the parallel benchmarks UI, and
only fixed the UI for sequential benchmarks. This commit fixes that.
