+++ app/apps/parallel_benchmarks.py
@@ -14,6 +14,7 @@ import pandas as pd
 import pandas.io.json as pdjson
 import seaborn as sns
 from apps import benchstruct
+from apps.utils import fmt_variant, unfmt_variant
 
 
 def app():
@@ -46,13 +47,6 @@ def app():
         # st.write(a)
         commit_variant_tuple_lst = [(x1, x2) for x1, x2 in zip(a[0], a[1])]
         return commit_variant_tuple_lst
-
-    def fmt_variant(commit, variant_lst):
-        fmt_variant_lst = [f"{commit};{v}" for v in variant_lst]
-        return fmt_variant_lst
-
-    def unfmt_variant(variant):
-        return variant.split(";", 1)
 
     def get_selected_values(n):
         lst = []
